+++ src/XUnity.AutoTranslator.Patcher/Patcher.cs
@@ -29,7 +29,7 @@ namespace XUnity.AutoTranslator.Patcher
       {
          get
          {
-            return "3.5.0";
+            return "3.6.0";
          }
       }