+++ src/XUnity.AutoTranslator.Plugin.Core/UI/GUIUtil.cs
@@ -21,7 +21,7 @@ namespace XUnity.AutoTranslator.Plugin.Core.UI
       {
          richText = false,
          margin = new RectOffset( GUI.skin.label.margin.left, GUI.skin.label.margin.right, 0, 0 ),
-         padding = new RectOffset( GUI.skin.label.padding.left, GUI.skin.label.padding.right, 0, 0 )
+         padding = new RectOffset( GUI.skin.label.padding.left, GUI.skin.label.padding.right, 2, 3 )
       };
 
       public static readonly GUIStyle LabelCenter = new GUIStyle( GUI.skin.label )