+++ src/XUnity.AutoTranslator.Plugin.Core/Utilities/CoroutineHelper.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections;
+﻿using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -10,5 +9,7 @@ namespace XUnity.AutoTranslator.Plugin.Core.Utilities
    internal static class CoroutineHelper
    {
       public static Coroutine Start( IEnumerator coroutine ) => AutoTranslationPlugin.Current.StartCoroutine( coroutine );
+
+      public static void Stop( Coroutine coroutine ) => AutoTranslationPlugin.Current.StopCoroutine( coroutine );
    }
 }